Got it looking kind of sweet, now for logic
Next is the 'total search' option, whereby the search bar will allow up
to 1000 gifs to be loaded into the allGifs array. Currently problem is
the loading of gifs on screen - masonry.js is working and arranging the
gifs nicely, but sometimes the gifs just don't load for a good 30
seconds.